Code refactoring to meet the REST coding naming convention. Adding additional validation to methods. Methods to return proper HTTP codes. Added update methods

+++ ajdu_restful_api/src/main/java/ajdu_restful_api/controller/CategoryRestController.java
@@ -4,12 +4,12 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import ajdu_restful_api.model.Category;
@@ -21,17 +21,12 @@ public class CategoryRestController {
 	@Autowired
 	private CategoryService categoryService;
 	
-	@GetMapping("/categories")
+	@RequestMapping(value="/categories", method=RequestMethod.GET,consumes=MediaType.APPLICATION_JSON_UTF8_VALUE)
 	public ResponseEntity<List<Category>> allCategories(){
 		return new ResponseEntity<List<Category>>(categoryService.findAllCategories(),HttpStatus.OK);
 	}
 	
-	@GetMapping("/category")
-	public ResponseEntity<Category> getCategory(@RequestParam int id) {
-		return new ResponseEntity<Category>(categoryService.findOneCategory(id),HttpStatus.OK);
-	}
-	
-	@RequestMapping(value="/addcategory", method=RequestMethod.POST)
+	@RequestMapping(value="/categories", method=RequestMethod.POST,consumes=MediaType.APPLICATION_JSON_UTF8_VALUE)
 	public ResponseEntity<Category> addCategory(@RequestBody Category cat) {
 		if(categoryService.findCategoryByName(cat.getName()) == null) {
 			categoryService.saveCategory(cat);
@@ -41,10 +36,32 @@ public class CategoryRestController {
 		}
 	}
 	
-	@GetMapping("/deletecategory")
-	public ResponseEntity<Category> deleteCategory(@RequestParam int id) {
-		categoryService.deleteCategory(id);
-		return new ResponseEntity<Category>(HttpStatus.OK);
+	@RequestMapping(value="/categories/{id}",method=RequestMethod.GET,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
+	public ResponseEntity<Category> getCategory(@PathVariable int id) {
+		if(categoryService.findOneCategory(id) != null)
+			return new ResponseEntity<Category>(categoryService.findOneCategory(id),HttpStatus.OK);
+		else return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);			
+	}
+	
+	
+	@RequestMapping(value="/categories/{id}",method=RequestMethod.DELETE,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
+	public ResponseEntity<Category> deleteCategory(@PathVariable int id) {
+		Category c = categoryService.findOneCategory(id); 
+		if(c != null) {
+			categoryService.deleteCategory(id);
+			return new ResponseEntity<Category>(HttpStatus.OK);
+		}
+		return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);
+	}
+	
+	@RequestMapping(value="/categories/{id}",method=RequestMethod.PUT,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
+	public ResponseEntity<Category> updateCategory(@PathVariable int id, @RequestBody Category category) {
+		Category c = categoryService.findOneCategory(id); 
+		if(c != null) {
+			c.setName(category.getName());
+			return new ResponseEntity<Category>(c,HttpStatus.OK);
+		}
+		return new ResponseEntity<Category>(HttpStatus.NOT_FOUND);
 	}
