+++ ajdu_restful_api/src/main/java/ajdu_restful_api/controller/TaskRestController.java
@@ -4,8 +4,12 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -22,26 +26,44 @@ public class TaskRestController {
 	@Autowired
 	ScheduleService scheduleService;
 	
-	@GetMapping("/task")
-	public Task findTask(@RequestParam int id) {
-		return taskService.findTask(id);
+	@RequestMapping(value="/tasks",method=RequestMethod.GET,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
+	public ResponseEntity<List<Task>> getAllTask() {
+			return new ResponseEntity<List<Task>>(taskService.findAll(),HttpStatus.OK);
 	}
 	
-	@GetMapping("/alltasks")
-	public List<Task> findAll(@RequestParam int scheduleID){
-		return taskService.findAllTaskBySchedule(scheduleID);
+	@RequestMapping(value="/tasks/{id}",method=RequestMethod.GET,produces=MediaType.APPLICATION_JSON_UTF8_VALUE)
+	public ResponseEntity<Task> getTask(@PathVariable int id) {
+		if(taskService.findTask(id) != null)
+			return new ResponseEntity<Task>(taskService.findTask(id),HttpStatus.OK);
+		else return new ResponseEntity<Task>(HttpStatus.NOT_FOUND);
 	}
 	
-	@GetMapping("/savetask")
+	@RequestMapping(value="/tasks",method=RequestMethod.POST,consumes=MediaType.APPLICATION_JSON_UTF8_VALUE)
 	public ResponseEntity<Task> saveTask(@RequestParam String name, @RequestParam int scheduleId) {
 		Task t = new Task(name, scheduleService.findSchedule(scheduleId));
 		taskService.saveTask(t);
 		return new ResponseEntity<Task>(t,HttpStatus.CREATED);
 	}
 	
-	@GetMapping("/deletetask")
-	public String deleteTask(@RequestParam int id){
-		taskService.deleteTask(id);
-		return "Task with id "+id+" has been removed.";
+	@RequestMapping(value="/tasks/{id}",method=RequestMethod.DELETE)
+	public ResponseEntity<Task> deleteTask(@PathVariable int id) {
+		if(taskService.findTask(id) != null) {
+			taskService.deleteTask(id);
+			return new ResponseEntity<Task>(HttpStatus.OK);
+		}
+		else return new ResponseEntity<Task>(HttpStatus.NOT_FOUND);
+	}
+	
+	@RequestMapping(value="/tasks/{id}",method=RequestMethod.PUT, consumes=MediaType.APPLICATION_JSON_UTF8_VALUE)
+	public ResponseEntity<Task> updateTask(@PathVariable int id, @RequestBody Task task) {
+		Task t = taskService.findTask(id);
+		if(t != null) {
+			t.setName(task.getName());
+			t.setDescription(task.getDescription());
+			t.setDueDate(task.getDueDate());
+			t.setStatus(task.getStatus());
+			return new ResponseEntity<Task>(t, HttpStatus.OK);
+		}
+		else return new ResponseEntity<Task>(HttpStatus.NOT_FOUND);
 	}
 }
